requirements.txt
pytesseract
google-generativeai
